refactor(workbench-cli): move the Brett API wrappers into services
