+++ packages/@sanity/workbench-cli/src/actions/deploy/deployConfig.ts
@@ -1,22 +1,14 @@
 import {basename, dirname} from 'node:path'
-import {PassThrough} from 'node:stream'
 import {styleText} from 'node:util'
 import {createGzip} from 'node:zlib'
 
-import {getGlobalCliClient, type Output, subdebug} from '@sanity/cli-core'
-import FormData from 'form-data'
+import {type Output, subdebug} from '@sanity/cli-core'
 import {pack} from 'tar-fs'
 
-import {APP_WORKBENCH_API_VERSION} from './apiVersion.js'
+import {createConfig, resolveSingletonInstallationId} from '../../services/installations.js'
 import {summarizeExposeGroup} from './buildExposes.js'
 
 const debug = subdebug('deploy')
-
-interface InstallationListItem {
-  id: string
-
-  application?: {slug?: string}
-}
 
 /**
  * The org's active installation for an app type, or `undefined` when none is
@@ -71,41 +63,8 @@ export async function deployConfig(options: {
 }): Promise<void> {
   const {appType, installationId, output, sourceDir, version} = options
   const tarball = pack(dirname(sourceDir), {entries: [basename(sourceDir)]}).pipe(createGzip())
-  const formData = new FormData()
-  formData.append('version', version)
-  formData.append('tarball', tarball, {
-    contentType: 'application/gzip',
-    filename: 'installation-config.tar.gz',
-  })
-
-  const client = await getGlobalCliClient({
-    apiVersion: APP_WORKBENCH_API_VERSION,
-    requireUser: true,
-  })
-  await client.request({
-    body: formData.pipe(new PassThrough()),
-    headers: formData.getHeaders(),
-    method: 'POST',
-    uri: `/installations/${installationId}/configs`,
-  })
+  await createConfig(installationId, {tarball, version})
 
   debug('Deployed config for app type: %s', appType)
   output.log(`\n🚀 ${styleText('bold', 'Success!')} Config deployed`)
-}
-
-/** The org's active singleton installation, matched on its slug. */
-async function resolveSingletonInstallationId(
-  organizationId: string,
-  slug: string,
-): Promise<string | undefined> {
-  const client = await getGlobalCliClient({
-    apiVersion: APP_WORKBENCH_API_VERSION,
-    requireUser: true,
-  })
-  // `limit=none` returns every installation in one response, no pagination.
-  const {data}: {data: InstallationListItem[]} = await client.request({
-    query: {limit: 'none', organizationId},
-    uri: '/installations',
-  })
-  return data.find((item) => item.application?.slug === slug)?.id
 }
