+++ packages/@sanity/workbench-cli/src/actions/deploy/deployWorkbenchApp.ts
@@ -1,141 +1,16 @@
 import {basename, dirname} from 'node:path'
-import {PassThrough} from 'node:stream'
-import {createGzip, type Gzip} from 'node:zlib'
+import {createGzip} from 'node:zlib'
 
-import {exitCodes, getGlobalCliClient, type Output} from '@sanity/cli-core'
+import {exitCodes, type Output} from '@sanity/cli-core'
 import {spinner} from '@sanity/cli-core/ux'
-import FormData from 'form-data'
 import {pack} from 'tar-fs'
 
-import {APP_WORKBENCH_API_VERSION} from './apiVersion.js'
-import {type BrettInterface} from './buildExposes.js'
-
-export type ApplicationType = 'coreApp' | 'studio'
-
-export interface Application {
-  id: string
-  organizationId: string
-  slug: string | null
-  title: string
-  type: ApplicationType
-}
-
-/** A studio workspace as Brett stores it. */
-export interface BrettWorkspace {
-  dataset: string
-  projectId: string
-
-  basePath?: string
-  icon?: string
-  name?: string
-  subtitle?: string
-  title?: string
-}
-
-export async function getApplication(applicationId: string): Promise<Application | null> {
-  const client = await getGlobalCliClient({
-    apiVersion: APP_WORKBENCH_API_VERSION,
-    requireUser: true,
-  })
-  try {
-    return await client.request({uri: `/applications/${applicationId}`})
-  } catch (err) {
-    if ((err as {statusCode?: number})?.statusCode === 404) return null
-    throw err
-  }
-}
-
-/** Create an application and its first deployment in one call. */
-export async function createApplication(options: {
-  interfaces: readonly BrettInterface[]
-  isSingleton?: boolean
-  organizationId: string
-  projectId?: string
-  slug: string
-  tarball: Gzip
-  title: string
-  type: ApplicationType
-  version: string
-  workspaces?: readonly BrettWorkspace[]
-}): Promise<Application> {
-  const {
-    interfaces,
-    isSingleton,
-    organizationId,
-    projectId,
-    slug,
-    tarball,
-    title,
-    type,
-    version,
-    workspaces,
-  } = options
-  const formData = new FormData()
-  formData.append('type', type)
-  formData.append('title', title)
-  formData.append('organizationId', organizationId)
-  formData.append('slug', slug)
-  if (isSingleton !== undefined) formData.append('isSingleton', String(isSingleton))
-  // Studio config is set once, at create — it's immutable on redeploy.
-  if (projectId) appendJson(formData, 'config', {studio: {projectId}})
-  appendDeploymentParts(formData, {interfaces, tarball, version, workspaces})
-  return request(`/applications`, formData)
-}
-
-/** Deploy a new active version to an existing application. */
-export async function createDeployment(options: {
-  applicationId: string
-  interfaces: readonly BrettInterface[]
-  isAutoUpdating: boolean
-  tarball: Gzip
-  version: string
-  workspaces?: readonly BrettWorkspace[]
-}): Promise<{id: string}> {
-  const {applicationId, interfaces, isAutoUpdating, tarball, version, workspaces} = options
-  const formData = new FormData()
-  formData.append('isAutoUpdating', isAutoUpdating.toString())
-  appendDeploymentParts(formData, {interfaces, tarball, version, workspaces})
-  return request(`/applications/${applicationId}/deployments`, formData)
-}
-
-function appendDeploymentParts(
-  formData: FormData,
-  {
-    interfaces,
-    tarball,
-    version,
-    workspaces,
-  }: {
-    interfaces: readonly BrettInterface[]
-    tarball: Gzip
-    version: string
-    workspaces?: readonly BrettWorkspace[]
-  },
-): void {
-  formData.append('version', version)
-  appendJson(formData, 'interfaces', interfaces)
-  // Studio-only — the server rejects a workspaces part on non-studio types.
-  if (workspaces?.length) appendJson(formData, 'workspaces', workspaces)
-  formData.append('tarball', tarball, {contentType: 'application/gzip', filename: 'app.tar.gz'})
-}
-
-/** Structured parts must arrive as JSON so the server parses them. */
-function appendJson(formData: FormData, name: string, value: unknown): void {
-  formData.append(name, JSON.stringify(value), {contentType: 'application/json'})
-}
-
-async function request<T>(uri: string, formData: FormData): Promise<T> {
-  const client = await getGlobalCliClient({
-    apiVersion: APP_WORKBENCH_API_VERSION,
-    requireUser: true,
-  })
-  return client.request({
-    body: formData.pipe(new PassThrough()),
-    headers: formData.getHeaders(),
-    method: 'POST',
-    uri,
-  })
-}
+import {
+  type BrettInterface,
+  type BrettWorkspace,
+  createApplication,
+  createDeployment,
+} from '../../services/applications.js'
 
 /**
  * Deploy a workbench coreApp through Brett: redeploy when `appId` is set,
